split out cli.py
