+++ source/autologic.py
@@ -1,12 +1,5 @@
-import click
-import csv
 import math
 import random
-import yaml
-
-from pathlib import Path
-from pydantic import BaseModel, Field, ValidationError
-from typing import Optional
 
 import utils
 from Event import Event
@@ -15,6 +8,7 @@ from Event import Event
 # TODO: refactor for sanity and flexibility
 #       this should be split out into separate functions and perhaps even modules
 #       left as-is for quick prototype development ahead of a pilot event
+#       this module is intended to house algorithms and algorithm-related functions
 # =============================================================================
 
 
@@ -29,89 +23,6 @@ def randomize_heats(event, number_of_heats):
     random.shuffle(categories)
     for i, c in enumerate(categories):
         c.set_heat(i % number_of_heats + 1)
-
-
-class Config(BaseModel):
-    axware_export_tsv: Path = Field(..., description="Path to AXWare export TSV file.")
-    member_attributes_csv: Path = Field(
-        ..., description="Path to member attribute CSV file."
-    )
-    number_of_heats: int = Field(
-        3, description="Number of heats to divide participants into."
-    )
-    number_of_stations: int = Field(
-        5, description="Number of worker stations for the course."
-    )
-    custom_assignments: dict[str | int, str | list[str]] = Field(
-        default_factory=dict,
-        description="A dictionary of member IDs to their custom role assignments.",
-    )
-    heat_size_parity: int = Field(
-        25, description="Smaller values enforce tighter heat size balance."
-    )
-    novice_size_parity: int = Field(
-        10, description="Smaller values enforce tighter novice balance across heats."
-    )
-    novice_denominator: int = Field(
-        3, description="Min instructors in heat = novices / denominator."
-    )
-    max_iterations: int = Field(
-        10000, description="Max number of attempts before giving up."
-    )
-
-    def validate_paths(self):
-        """Ensure all paths exist and are files."""
-        for path_attr in ["axware_export_tsv", "member_attributes_csv"]:
-            p = getattr(self, path_attr)
-            if not p.exists():
-                raise FileNotFoundError(f"{path_attr} does not exist: {p}")
-            if not p.is_file():
-                raise ValueError(f"{path_attr} is not a file: {p}")
-
-
-def load_config(ctx, param, value: Path) -> Config:
-    if value is None:
-        return None
-    try:
-        with open(value, "r") as f:
-            data = yaml.safe_load(f)
-        config = Config(**data)
-        config.validate_paths()
-        return config
-    except (ValidationError, FileNotFoundError, ValueError) as e:
-        raise click.BadParameter(f"Invalid config: {e}")
-    except Exception as e:
-        raise click.BadParameter(f"Failed to load config: {e}")
-
-
-@click.command(context_settings={"max_content_width": 120})
-@click.option(
-    "--config",
-    type=click.Path(exists=True, dir_okay=False, readable=True, path_type=Path),
-    callback=load_config,
-    required=False,
-    help="Path to event configuration file.",
-)
-@click.option(
-    "--to-pdf",
-    type=click.Path(exists=True, dir_okay=False, readable=True, path_type=Path),
-    required=False,
-    help="Path to autologic-export.csv file for conversion to PDF. Useful for making manual tweaks and then printing.",
-)
-def cli(config: Optional[dict], to_pdf: Optional[Path]):
-
-    if not config and not to_pdf:
-        raise click.UsageError("You must provide either --config or --to-pdf.")
-    if config and to_pdf:
-        raise click.UsageError("Only one of --config or --to-pdf can be used.")
-
-    if config:
-        main(**config.model_dump())
-    elif to_pdf:
-        with open(to_pdf, newline="", encoding="utf-8-sig") as file:
-            rows = csv.DictReader(file)
-            utils.autologic_rows_to_pdf(rows)
-            print()
 
 
 def main(
@@ -303,7 +214,3 @@ def main(
     utils.autologic_rows_to_csv(rows)
     utils.autologic_rows_to_pdf(rows)
     print()
-
-
-if __name__ == "__main__":
-    cli()
